+++ internal/pkg/model/court.go
@@ -6,9 +6,9 @@ import (
 )
 
 type Court struct {
-	ID          int64     `json:"id" gorm:"primary_key;AUTO_INCREMENT"`
+	ID          int32     `json:"id" gorm:"primary_key;AUTO_INCREMENT"`
 	CourtName   string    `json:"court_name" gorm:"column:court_name;type:varchar(255);default:'';comment:'场馆名称'"`
-	VenueId     int64     `json:"venue_id" gorm:"column:venue_id;not null;comment:'场馆id'"`
+	VenueId     int32     `json:"venue_id" gorm:"column:venue_id;not null;comment:'场馆id'"`
 	CreatedTime time.Time `json:"created_time" gorm:"column:created_time;type:datetime;not null;default:CURRENT_TIMESTAMP;comment:'创建时间'"`
 	UpdatedTime time.Time `json:"updated_time" gorm:"column:updated_time;type:datetime;not null;default:CURRENT_TIMESTAMP;comment:'更新时间'"`
 }
